+++ src/main/kotlin/me/notsmatch/kyoshubot/command/DropCommand.kt
@@ -23,8 +23,9 @@ class DropCommand(val boshuService: BoshuService, val settingsService: GuildSett
 
     override fun execute(event: CommandEvent?) {
         event?.apply {
-            val settings = settingsService.getGuildSettings(guild.idLong)
 
+            val settings = settingsService.getGuildSettings(guild.idLong)
+            if(settings.banned)return reply("This server has been banned.")
             if (settings.getCommandOption("d") == null || !settings.getCommandOption("d")!!.visibility) {
                 event.message.delete().complete()
             }
