+++ src/main/kotlin/me/notsmatch/kyoshubot/command/EndCommand.kt
@@ -18,8 +18,9 @@ class EndCommand(val boshuService: BoshuService, val settingsService: GuildSetti
 
     override fun execute(event: CommandEvent?) {
         event?.apply {
-            val settings = settingsService.getGuildSettings(guild.idLong)
 
+            val settings = settingsService.getGuildSettings(guild.idLong)
+            if(settings.banned)return reply("This server has been banned.")
             if (settings.getCommandOption("end") == null || !settings.getCommandOption("end")!!.visibility) {
                 event.message.delete().complete()
             }
