+++ src/main/kotlin/me/notsmatch/kyoshubot/command/ResendCommand.kt
@@ -20,7 +20,7 @@ class ResendCommand (val boshuService: BoshuService,  val settingsService: Guild
         event?.apply {
 
             val settings = settingsService.getGuildSettings(guild.idLong)
-
+            if(settings.banned)return reply("This server has been banned.")
             if (settings.getCommandOption("resend") == null || !settings.getCommandOption("resend")!!.visibility) {
                 event.message.delete().complete()
             }
